TF-933 Update new position all button `QuickSearchFilter` on Web Desktop
(cherry picked from commit 0fb78cef5ea424700eec48dc1a11914183804a10)
