+++ lib/features/mailbox_dashboard/presentation/mailbox_dashboard_view_web.dart
@@ -13,6 +13,7 @@ import 'package:tmail_ui_user/features/mailbox_dashboard/presentation/action/das
 import 'package:tmail_ui_user/features/mailbox_dashboard/presentation/base_mailbox_dashboard_view.dart';
 import 'package:tmail_ui_user/features/mailbox_dashboard/presentation/controller/search_controller.dart';
 import 'package:tmail_ui_user/features/mailbox_dashboard/presentation/model/composer_overlay_state.dart';
+import 'package:tmail_ui_user/features/mailbox_dashboard/presentation/model/search/email_receive_time_type.dart';
 import 'package:tmail_ui_user/features/mailbox_dashboard/presentation/model/search/quick_search_filter.dart';
 import 'package:tmail_ui_user/features/mailbox_dashboard/presentation/widgets/advanced_search/advanced_search_filter_overlay.dart';
 import 'package:tmail_ui_user/features/mailbox_dashboard/presentation/widgets/advanced_search/icon_open_advanced_search_widget.dart';
@@ -95,17 +96,8 @@ class MailboxDashBoardView extends BaseMailboxDashBoardView {
                     children: [
                       SizedBox(child: MailboxView(), width: responsiveUtils.defaultSizeMenu),
                       Expanded(child: Column(children: [
-                        Obx(() {
-                          if (controller.vacationResponse.value?.vacationResponderIsValid == true) {
-                            return VacationNotificationMessageWidget(
-                                margin: const EdgeInsets.only(top: 16, right: 16),
-                                vacationResponse: controller.vacationResponse.value!,
-                                actionGotoVacationSetting: () => controller.goToVacationSetting(),
-                                actionEndNow: () => controller.disableVacationResponder());
-                          } else {
-                            return const SizedBox.shrink();
-                          }
-                        }),
+                        _buildVacationNotificationMessage(context),
+                        _buildListButtonQuickSearchFilter(context),
                         _buildMarkAsMailboxReadLoading(context),
                         Expanded(child: Obx(() {
                           switch(controller.routePath.value) {
@@ -680,5 +672,165 @@ class MailboxDashBoardView extends BaseMailboxDashBoardView {
         return const SizedBox.shrink();
       }
     });
+  }
+
+  Widget _buildVacationNotificationMessage(BuildContext context) {
+    return Obx(() {
+      if (controller.vacationResponse.value?.vacationResponderIsValid == true) {
+        return VacationNotificationMessageWidget(
+            margin: const EdgeInsets.only(top: 16, right: 16),
+            vacationResponse: controller.vacationResponse.value!,
+            actionGotoVacationSetting: () => controller.goToVacationSetting(),
+            actionEndNow: () => controller.disableVacationResponder());
+      } else {
+        return const SizedBox.shrink();
+      }
+    });
+  }
+
+  bool supportListButtonQuickSearchFilter(BuildContext context) {
+    return controller.searchController.isSearchActive()
+        && controller.searchController.searchIsActive.isTrue
+        && responsiveUtils.isWebDesktop(context);
+  }
+
+  Widget _buildListButtonQuickSearchFilter(BuildContext context) {
+    return Obx(() {
+      if (supportListButtonQuickSearchFilter(context)) {
+        return Padding(
+          padding: const EdgeInsets.only(right: 16, top: 16),
+          child: Row(
+              children: QuickSearchFilter.values
+                  .map((filter) => _buildQuickSearchFilterButton(context, filter))
+                  .toList()),
+        );
+      } else {
+        return const SizedBox.shrink();
+      }
+    });
+  }
+
+  Widget _buildQuickSearchFilterButton(
+      BuildContext context,
+      QuickSearchFilter filter
+  ) {
+    return Obx(() {
+      final quickSearchFilterSelected = controller.checkQuickSearchFilterSelected(
+        quickSearchFilter: filter,
+      );
+
+      return Padding(
+        padding: const EdgeInsets.only(right: 8),
+        child: InkWell(
+          onTap: () {
+            if (filter != QuickSearchFilter.last7Days) {
+              controller.selectQuickSearchFilterAction(filter);
+            }
+          },
+          onTapDown: (detail) {
+            if (filter == QuickSearchFilter.last7Days) {
+              final screenSize = MediaQuery.of(context).size;
+              final offset = detail.globalPosition;
+              final position = RelativeRect.fromLTRB(
+                offset.dx,
+                offset.dy,
+                screenSize.width - offset.dx,
+                screenSize.height - offset.dy,
+              );
+              controller.openPopupMenuAction(context, position,
+                  popupMenuEmailReceiveTimeType(context,
+                      controller.searchController.emailReceiveTimeType.value,
+                          (receiveTime) => controller.selectReceiveTimeQuickSearchFilter(receiveTime)));
+            }
+          },
+          borderRadius: const BorderRadius.all(Radius.circular(10)),
+          child: Container(
+              decoration: BoxDecoration(
+                  borderRadius: BorderRadius.circular(10),
+                  color: filter.getBackgroundColor(quickSearchFilterSelected: quickSearchFilterSelected)),
+              padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 8),
+              child: Row(mainAxisAlignment: MainAxisAlignment.center, children: [
+                SvgPicture.asset(
+                    filter.getIcon(
+                        imagePaths,
+                        quickSearchFilterSelected: quickSearchFilterSelected),
+                    width: 16,
+                    height: 16,
+                    fit: BoxFit.fill),
+                const SizedBox(width: 4),
+                Text(
+                  filter.getTitle(
+                      context,
+                      receiveTimeType: controller.searchController.emailReceiveTimeType.value),
+                  maxLines: 1,
+                  overflow: TextOverflow.ellipsis,
+                  style: filter.getTextStyle(
+                      quickSearchFilterSelected: quickSearchFilterSelected),
+                ),
+                if (filter == QuickSearchFilter.last7Days)
+                  ... [
+                    const SizedBox(width: 4),
+                    SvgPicture.asset(
+                        imagePaths.icChevronDown,
+                        width: 16,
+                        height: 16,
+                        fit: BoxFit.fill),
+                  ]
+              ])),
+        ),
+      );
+    });
+  }
+
+  List<PopupMenuEntry> popupMenuEmailReceiveTimeType(
+      BuildContext context,
+      EmailReceiveTimeType? receiveTimeSelected,
+      Function(EmailReceiveTimeType?)? onCallBack
+  ) {
+    return EmailReceiveTimeType.values
+        .map((timeType) => PopupMenuItem(
+            padding: EdgeInsets.zero,
+            child: _receiveTimeTileAction(
+                context,
+                receiveTimeSelected,
+                timeType,
+                onCallBack)))
+        .toList();
+  }
+
+  Widget _receiveTimeTileAction(
+      BuildContext context,
+      EmailReceiveTimeType? receiveTimeSelected,
+      EmailReceiveTimeType receiveTimeType,
+      Function(EmailReceiveTimeType?)? onCallBack
+  ) {
+    return InkWell(
+        onTap: () => onCallBack?.call(receiveTimeType == receiveTimeSelected
+            ? null
+            : receiveTimeType),
+        child: Padding(
+          padding: const EdgeInsets.symmetric(horizontal: 24, vertical: 16),
+          child: SizedBox(
+              width: 320,
+              child: Row(children: [
+                Expanded(child: Text(
+                    receiveTimeType.getTitle(context),
+                    style: const TextStyle(
+                        fontSize: 17,
+                        color: Colors.black,
+                        fontWeight: FontWeight.normal))),
+                if (receiveTimeType == receiveTimeSelected)
+                  ...[
+                    const SizedBox(width: 12),
+                    SvgPicture.asset(
+                        imagePaths.icFilterSelected,
+                        width: 24,
+                        height: 24,
+                        fit: BoxFit.fill),
+                  ]
+              ])
+          ),
+        )
+    );
   }
 }
